Changed message packaging to make more sense.
If we receive a request packet, we send a response. We don't respond to request packets. The idea behind this is that the server only sends success/failure responses to the client, but not to other servers. Client/server communication is request response, server/server is messge only. Finishing transfers and propogating snapshots are server/server communication, and should use messages; everything else (so far) should be request/response.